+++ tm.py
@@ -32,6 +32,7 @@ web.isHardened = True
 web.sanitizesInput = False
 web.encodesOutput = True
 web.authorizesSource = False
+web.sourceFiles = ["pytm/json.py", "docs/template.md"]
 
 db = Datastore("SQL Database")
 db.OS = "CentOS"
@@ -44,6 +45,7 @@ db.levels = [2]
 
 secretDb = Datastore("Real Identity Database")
 secretDb.OS = "CentOS"
+secretDb.sourceFiles = ["pytm/pytm.py"]
 secretDb.isHardened = True
 secretDb.inBoundary = server_db
 secretDb.isSQL = True