+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ## Bug fixes
 
 - Fixed printing the data attribute in reports [#123](https://github.com/izar/pytm/pull/123)
+- Added a markdown file with threats [#126](https://github.com/izar/pytm/pull/126)
 - Fixed drawing nested boudnaries [#117](https://github.com/izar/pytm/pull/117)
 
 # 1.1.2